Create  Map, reduce and filter
Takes integers, names, and numbers as the parameters.
a. Use map() to return the list of the squares of each number.
b. Use filter() to return the list with only those names that have less than or equal to seven letters.
c. Use reduce() to return the product of elements of list numbers.